+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - NumericRange: Fixed null value display bug
 - Tour: Added class 'tour-active-element' for active element to customize styles
 - FileDrop, DropZone: Added loadingText prop
+- ToolTip: Updated styles
 
 
 ## [0.34.0] - 2020-11-09